enhanced/requirements_enhanced.txt
# PowerTrader Enhanced - Requirements

# Core dependencies
python>=3.10

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP clients
requests>=2.31.0
httpx>=0.25.0

# Crypto exchange APIs
kucoin-python>=2.1.3
ccxt>=4.1.0  # Multi-exchange support

# Cryptography
PyNaCl>=1.5.0
cryptography>=41.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.9  # PostgreSQL
asyncpg>=0.29.0  # Async PostgreSQL

# Caching & message queue
redis>=5.0.0
hiredis>=2.2.3

# Data processing
numpy>=1.24.0
pandas>=2.1.0
scipy>=1.11.0

# Async utilities
aiofiles>=23.2.0
aiocache>=0.12.0

# System monitoring
psutil>=5.9.0

# Logging
structlog>=23.2.0
python-json-logger>=2.0.0

# Configuration
python-dotenv>=1.0.0

# CLI utilities
click>=8.1.0
rich>=13.7.0
typer>=0.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
httpx-mock>=0.9.0

# Code quality
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
pylint>=3.0.0
mypy>=1.7.0
pre-commit>=3.5.0

# Type stubs
types-requests>=2.31.0
types-redis>=4.6.0

# Development
ipython>=8.17.0
jupyter>=1.0.0

# Monitoring & metrics
prometheus-client>=0.19.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0

# Alerting
discord.py>=2.3.0
python-telegram-bot>=20.7
slack-sdk>=3.26.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.4.0

# Charting (for dashboard)
matplotlib>=3.8.0
plotly>=5.18.0

# Date/time
python-dateutil>=2.8.0
pytz>=2023.3

# Environment
colorama>=0.4.6

# Performance
orjson>=3.9.0  # Fast JSON
uvloop>=0.19.0  # Fast event loop (Unix only)

# Security
python-jose[cryptography]>=3.3.0  # JWT
passlib[bcrypt]>=1.7.4

# Rate limiting
slowapi>=0.1.9

# CORS
python-multipart>=0.0.6

# Health checks
py-healthcheck>=1.10.0
